+++ test/runner/browser/runner_test.dart
@@ -41,6 +41,32 @@ void main() {
       test.shouldExit(1);
     }, tags: 'chrome');
 
+    test("a test file has import errors", () {
+      // Bad imports cause errors only at runtime when running in content-shell
+      // or Dartium. The runner will connect to the browser but will never
+      // receive the test suite to deserialize.
+      d
+          .file(
+              "test.dart",
+              """
+              import 'package:test/test.dart';
+              import 'blah' as bad_library;
+
+              void main() {
+                bad_library.main();
+                test("failure", () => throw new TestFailure("oh no"));
+              }
+              """)
+          .create();
+      var test = runTest(["-p", "content-shell", "test.dart"]);
+
+      test.stdout.expect(containsInOrder([
+        "Timed out while loading the test suite.",
+        "It's likely that there's a missing import or syntax error."
+      ]));
+      test.shouldExit(1);
+    }, tags: 'content-shell');
+
     test("a test file throws", () {
       d.file("test.dart", "void main() => throw 'oh no';").create();