+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 * Only print browser load errors once per browser.
 
+* Gracefully time out when attempting to deserialize a test suite.
+
 ## 0.12.20+13
 
 * Upgrade to package:matcher 0.12.1